feat: correctly display public and shielded values in rows
